requirements.txt
openai
openai-whisper
librosa
transformers
torch